Fix regression and move procurement order code in other module.
