+++ purchase_request_to_rfq/models/__init__.py
@@ -4,5 +4,4 @@
 
 from . import purchase_request
 from . import purchase_order
-from . import procurement_order
 from . import stock
